@@ -1,10 +1,10 @@
 {
   "name": "@extension-packages/ag-grid-angular",
-  "version": "0.1.5",
+  "version": "0.2.0",
   "peerDependencies": {
-    "@angular/common": ">=17.0.0",
-    "@angular/core": ">=17.0.0",
-    "ag-grid-angular": ">=31.0.0"
+    "@angular/common": ">=18.0.0",
+    "@angular/core": ">=18.0.0",
+    "ag-grid-angular": ">=33.0.0"
   },
   "dependencies": {
     "tslib": "^2.3.0"
