@@ -1,22 +1,34 @@
-import { Component, Input } from '@angular/core';
+import { Component, input } from '@angular/core';
 
 @Component({
   selector: 'ag-grid-toolbar-action',
-  templateUrl: './ag-grid-toolbar-action.component.html',
-  styleUrls: ['./ag-grid-toolbar-action.component.scss'],
+  standalone: true,
+  styleUrls: ['./ag-grid-toolbar-action.component.css'],
+  template: `
+    <span
+      class="toolbar-action"
+      [class.disabled]="disabled()"
+      [id]="id()"
+      [class]="iconFont()"
+      [style.color]="disabled() ? '' : color()"
+      [title]="tooltip()"
+    >
+      {{ icon() }}<ng-content></ng-content>
+    </span>
+  `,
 })
 export class AgGridToolbarActionComponent {
-  @Input() color?: string;
-  @Input() disabled?: boolean;
-  /** A font class for the icon. Default: `ag-icon`.
+  readonly color = input<string>();
+  readonly disabled = input<boolean>();
+  /** A font class for the icon. Default: `material-icons`.
    *
    * The font needs to be included in your html and css.
    *
    * **Example font setup:**
    * - https://developers.google.com/fonts/docs/material_icons#icon_font_for_the_web */
-  @Input() iconFont = 'ag-icon';
+  readonly iconFont = input('material-icons');
   /** ID for toolbar action, used to set html id */
-  @Input() id?: string;
+  readonly id = input<string>();
   /** Icon name. Example `home`
    *
    *
@@ -36,6 +48,6 @@ export class AgGridToolbarActionComponent {
    * And set `iconFont: 'material-icons'`
    * ```
    */
-  @Input() icon?: string;
-  @Input() tooltip?: string;
+  readonly icon = input<string>();
+  readonly tooltip = input<string>();
 }
