@@ -1,4 +1,3 @@
-import { GridOptions } from 'ag-grid-community';
 import {
   componentWrapperDecorator,
   moduleMetadata,
@@ -7,12 +6,28 @@ import {
 } from '@storybook/angular';
 import { AgGridModule } from 'ag-grid-angular';
 import {
-  AgGridExtensionModule,
+  AllCommunityModule,
+  GridOptions,
+  ModuleRegistry,
+  themeAlpine,
+  themeBalham,
+  themeMaterial,
+  themeQuartz,
+} from 'ag-grid-community';
+import {
   AgGridExtensionComponent,
   AgGridToolbarAction,
   actionsSets,
 } from '../../public-api';
 import { gridOptions } from '../../stories/story-helpers/grid-options';
+ModuleRegistry.registerModules([AllCommunityModule]);
+
+const themeMap = {
+  alpine: themeAlpine,
+  balham: themeBalham,
+  material: themeMaterial,
+  quartz: themeQuartz,
+};
 
 const actions: AgGridToolbarAction[] = [
   {
@@ -23,7 +38,7 @@ const actions: AgGridToolbarAction[] = [
       api.flashCells({ rowNodes: transaction?.add });
     },
     color: 'seagreen',
-    icon: 'plus',
+    icon: 'add',
     tooltip: 'Add new row',
   },
   {
@@ -32,7 +47,7 @@ const actions: AgGridToolbarAction[] = [
     },
     color: 'dodgerblue',
     disabled: true,
-    icon: 'eye',
+    icon: 'visibility',
     tooltip: 'Edit',
     tooltipDisabled: 'Select a row to edit',
   },
@@ -44,7 +59,7 @@ const actions: AgGridToolbarAction[] = [
     },
     color: 'crimson',
     disabled: true,
-    icon: 'cancel',
+    icon: 'delete',
     tooltip: 'Delete selected row(s)',
     tooltipDisabled: 'Select row(s) to delete',
   },
@@ -68,13 +83,14 @@ const meta: Meta<AgGridExtensionComponent> = {
       (story) => `<div style="height: 500px;">${story}</div>`,
     ),
     moduleMetadata({
-      imports: [AgGridModule, AgGridExtensionModule],
+      imports: [AgGridModule],
     }),
   ],
   render: (args) => {
     return {
       props: {
         ...args,
+        themeMap,
         gridOptions: {
           ...gridOptions,
           ...(args as any).extraGridOptions,
@@ -82,14 +98,13 @@ const meta: Meta<AgGridExtensionComponent> = {
       },
       template: `
     <ag-grid-extension
-      [class]="agTheme"
       [actions]="actions"
       [disableSearch]="disableSearch"
       [debounceSearch]="debounceSearch"
       [enableSearch]="enableSearch"
       [placeholderSearch]="placeholderSearch">
       
-      <ag-grid-angular [gridOptions]="gridOptions"></ag-grid-angular>
+      <ag-grid-angular [gridOptions]="gridOptions" [theme]="themeMap[agTheme]"></ag-grid-angular>
     </ag-grid-extension>
     `,
     };
@@ -102,7 +117,7 @@ type Story = StoryObj<
   AgGridExtensionComponent & { extraGridOptions?: GridOptions }
 >;
 
-export const SeachAndActions: Story = {
+export const SearchAndActions: Story = {
   args: {
     ...defaults,
     actions: actions,
@@ -117,11 +132,11 @@ export const ContentProjections: Story = {
   render: (args) => ({
     props: {
       ...args,
+      themeMap,
       gridOptions,
     },
     template: `
     <ag-grid-extension
-      [class]="agTheme"
       [actions]="null"
       [disableSearch]="disableSearch"
       [debounceSearch]="debounceSearch"
@@ -133,53 +148,53 @@ export const ContentProjections: Story = {
       <div actionsLeft>actionsLeft &nbsp;</div>
       <div actionsRight>actionsRight &nbsp;</div>
 
-      <ag-grid-angular [gridOptions]="gridOptions"></ag-grid-angular>
+      <ag-grid-angular [gridOptions]="gridOptions" [theme]="themeMap[agTheme]"></ag-grid-angular>
     </ag-grid-extension>
     `,
   }),
 };
 
-export const UpdateActionOnClick: Story = {
-  args: {
-    ...defaults,
-    actions: [
-      {
-        icon: 'radio-button-off',
-        clickFn: (params) => {
-          const { action } = params;
-          if (action.icon === 'radio-button-off') {
-            action.icon = 'radio-button-on';
-            action.color = 'seagreen';
-          } else {
-            action.icon = 'radio-button-off';
-            action.color = '';
-          }
-          action.tooltip = `Action icon changed to ${action.icon}`;
-        },
-      },
-      {},
-      {
-        icon: 'checkbox-unchecked',
-        clickFn: (params) => {
-          const { action } = params;
-          if (action.icon === 'checkbox-unchecked') {
-            action.icon = 'checkbox-checked';
-            action.color = 'seagreen';
-          } else if (action.icon === 'checkbox-checked') {
-            action.icon = 'checkbox-indeterminate';
-            action.color = 'orange';
-          } else {
-            action.icon = 'checkbox-unchecked';
-            action.color = '';
-          }
-          action.tooltip = `Action icon changed to ${action.icon}`;
-        },
-      },
-    ],
-    extraGridOptions: {
-      onFirstDataRendered: (params) => {
-        params.api.getDisplayedRowAtIndex(0)?.setSelected(true);
-      },
-    },
-  },
-};
+// export const UpdateActionOnClick: Story = {
+//   args: {
+//     ...defaults,
+//     actions: [
+//       {
+//         icon: 'radio-button-off',
+//         clickFn: (params) => {
+//           const { action } = params;
+//           if (action.icon === 'radio-button-off') {
+//             action.icon = 'radio-button-on';
+//             action.color = 'seagreen';
+//           } else {
+//             action.icon = 'radio-button-off';
+//             action.color = '';
+//           }
+//           action.tooltip = `Action icon changed to ${action.icon}`;
+//         },
+//       },
+//       {},
+//       {
+//         icon: 'checkbox-unchecked',
+//         clickFn: (params) => {
+//           const { action } = params;
+//           if (action.icon === 'checkbox-unchecked') {
+//             action.icon = 'checkbox-checked';
+//             action.color = 'seagreen';
+//           } else if (action.icon === 'checkbox-checked') {
+//             action.icon = 'checkbox-indeterminate';
+//             action.color = 'orange';
+//           } else {
+//             action.icon = 'checkbox-unchecked';
+//             action.color = '';
+//           }
+//           action.tooltip = `Action icon changed to ${action.icon}`;
+//         },
+//       },
+//     ],
+//     extraGridOptions: {
+//       onFirstDataRendered: (params) => {
+//         params.api.getDisplayedRowAtIndex(0)?.setSelected(true);
+//       },
+//     },
+//   },
+// };
