@@ -2,8 +2,7 @@
  * Public API Surface of ag-grid-angular
  */
 
-export * from './lib/ag-grid-extension/ag-grid-extension.module';
-export * from './lib/ag-grid-extension/ag-grid-extension.component';
+export * from './lib/ag-grid-extension/components/ag-grid-extension/ag-grid-extension.component';
 export * from './lib/ag-grid-extension/components/ag-grid-toolbar-action/ag-grid-toolbar-action.component';
 export * from './lib/ag-grid-extension/interfaces/ag-grid-context';
 export * from './lib/ag-grid-extension/interfaces/ag-grid-toolbar-action';
