@@ -6,24 +6,14 @@ import { AgGridContext } from '../../public-api';
 const datasourceService = new DatasourceService();
 const gridOptionsService = new GridOptionsService();
 export const gridOptions = gridOptionsService.clientSide();
-gridOptions.rowSelection = 'multiple';
-gridOptions.columnDefs = [
-  {
-    checkboxSelection: true,
-    headerCheckboxSelection: true,
-    headerCheckboxSelectionFilteredOnly: true,
-    minWidth: 35,
-    suppressCellFlash: true,
-    width: 50,
-  },
-  ...gridOptionsService.getColDefs(),
-];
+gridOptions.rowSelection = { mode: 'multiRow' };
+gridOptions.columnDefs = [...gridOptionsService.getColDefs()];
 gridOptions.rowData = datasourceService.generateItems(15);
 gridOptions.onSelectionChanged = (event: AgGridEvent<any, AgGridContext>) => {
   const { api, context } = event;
   const nodes = api.getSelectedNodes();
-  const deleteAction = context.actions?.find(({ icon }) => icon === 'cancel');
-  const editAction = context.actions?.find(({ icon }) => icon === 'eye');
+  const deleteAction = context.actions?.find(({ icon }) => icon === 'delete');
+  const editAction = context.actions?.find(({ icon }) => icon === 'visibility');
   if (deleteAction) {
     deleteAction.disabled = nodes.length === 0;
   }
